pyproject.toml
[project]
name = "ibis-widget"
version = "0.1.0"
description = "Jupyterwidget for exploring ibis tables"
authors = [
    {name = "Nick Crews", email = "nicholas.b.crews@gmail.com"},
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "anywidget>=0.9.13",
    "ibis-framework>=9.0.0",
]

[tool.pdm.dev-dependencies]
dev = [
    "ipykernel",
    "duckdb",
]
lint = ["ruff"]
test = ["pytest"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
